+++ lib/src/service/entities/tag.dart
@@ -27,7 +27,7 @@ class Tag with _$Tag {
     @JsonKey(name: 'history') List<UsageStatistics>? histories,
 
     /// Whether the current token’s authorized user is following this tag.
-    bool? isFollowing,
+    @JsonKey(name: 'following') bool? isFollowing,
   }) = _Tag;
 
   factory Tag.fromJson(Map<String, Object?> json) => _$TagFromJson(json);